project/myproject/quickstart/serializers.py
from rest_framework import serializers
from quickstart.models import Post


class PostSerializer(serializers.ModelSerializer):

    class Meta:
        model = Post
        fields = ('id', 'author', 'text', 'created', 'updated')