project/myproject/requirements.txt
Django==3.1.4
gunicorn==20.0.4
djangorestframework==3.12.2